Доработал способ получения вакансий по api hh.ru, указав 10 понравившихся мне работадателей.
